etl/requirements.txt
fastapi
psycopg2-binary
numpy
uvicorn
